Cast to uint64

+++ src/rpcblockchain.cpp
@@ -84,8 +84,8 @@ Object blockToJSON(const CBlock& block, const CBlockIndex* blockindex, bool txDe
     result.push_back(Pair("bits", strprintf("%08x", block.nBits)));
     result.push_back(Pair("difficulty", GetDifficulty(blockindex)));
     result.push_back(Pair("chainwork", blockindex->nChainWork.GetHex()));
-    result.push_back(Pair("startlocation", (uint32_t)block.nStartLocation));
-    result.push_back(Pair("finalcalculation", (uint32_t)block.nFinalCalculation));
+    result.push_back(Pair("startlocation", (uint64_t)block.nStartLocation));
+    result.push_back(Pair("finalcalculation", (uint64_t)block.nFinalCalculation));
     
     if (blockindex->pprev)
         result.push_back(Pair("previousblockhash", blockindex->pprev->GetBlockHash().GetHex()));
